moving things to different locations
