+++ src/api/digit.py
@@ -1,4 +1,4 @@
-from src.resources.parent_resource_concepts import ContentResource
+from src.api.parent_resource_concepts import ContentResource
 from flask_restful import reqparse, request
 from flask import jsonify
 import logging as lg
