+++ src/modules/time/time.py
@@ -11,6 +11,8 @@ import logging as lg
 class TimePlugin(ConfigLoader, BoardLinker):
     # TODO: get the board linker to link outside of the class for all once \
     # TODO: and then populate classes like this with (classwrapper???)
+
+    # TODO: move ConfigLoader and other parents to common plugin parent
     def __init__(self, board, config_file: str):
         ConfigLoader.__init__(self, config_file)
         # TODO: some better dynamic linking
